Reorder IncrementalCompilationTests and extract OutputFileMapCreator
